# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Документація API для блогу",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Post Title"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "This is the content of the post."
                                            },
                                            "userId": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "example": "2024-10-10 18:00:00"
                                            },
                                            "updatedAt": {
                                                "type": "string",
                                                "example": "2024-10-10 18:00:00"
                                            },
                                            "commentsCount": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Post Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "This is the content of the post."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Post Title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This is the content of the post."
                                        },
                                        "userId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        },
                                        "updatedAt": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a specific post",
                "operationId": "b0f9d92879218f68a376b6edc6467c67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post details with comments and their count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Post Title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This is the content of the post."
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        },
                                        "comments_count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "comments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "This is a comment."
                                                    },
                                                    "post_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-10-10 18:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-10-10 18:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a specific post",
                "operationId": "929614537d76a081a92170a41ed3bdb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Post Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content of the post."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated Post Title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Updated content of the post."
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-10-10 18:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a specific post",
                "operationId": "fad1726bd24cd459fc1525b67985df39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/user": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts of the authenticated user with comments count",
                "operationId": "e22efe7d968222398259fc79c8ffd41b",
                "responses": {
                    "200": {
                        "description": "List of user posts with comments count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Post Title"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "This is the content of the post."
                                            },
                                            "userId": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "example": "2024-10-10 18:00:00"
                                            },
                                            "updatedAt": {
                                                "type": "string",
                                                "example": "2024-10-10 18:00:00"
                                            },
                                            "commentsCount": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "comments": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "content": {
                                                            "type": "string",
                                                            "example": "Nice post!"
                                                        },
                                                        "userId": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "createdAt": {
                                                            "type": "string",
                                                            "example": "2024-10-12 09:00:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}